Fix propType console errors in file list
These props were never required, I just typed them wrong. The build is converting typescript props into runtime React propTypes, which is causing errors in the console when this prop is not provided. Scrolling seems to be working fine in the file list; we're not obsessively controlling its scroll behavior the way we do for the graph because we're not scroll-syncing multiple Lists, so there's no post-React-update bug to fix there.